feat: implement builder
