+++ crates/batch-submitter/src/lib.rs
@@ -1,5 +1,4 @@
-pub mod context;
+pub mod builder;
 pub mod error;
-pub mod writer;
 
 pub use error::BatchSubmitterError;
